Rename file and classes, Update study mode widget
