+++ rca/static_src/sass/components/_toggle-switch.scss
@@ -1,4 +1,4 @@
-.toggle-switch {
+.mode-checkbox {
     $root: &;
     display: flex;
     flex-direction: row;
